refactor on login
