+++ app/src/main/java/org/ole/planet/myplanet/ui/sync/LoginActivity.kt
@@ -1,26 +1,33 @@
 package org.ole.planet.myplanet.ui.sync
 
-import android.os.Build
-import android.os.Bundle
-import android.view.View
-import android.view.ViewGroup
+import android.content.*
+import android.graphics.drawable.AnimationDrawable
+import android.os.*
+import android.text.*
+import android.view.*
+import android.view.inputmethod.EditorInfo
+import android.widget.*
 import androidx.activity.OnBackPressedCallback
 import androidx.annotation.RequiresApi
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import androidx.appcompat.app.AlertDialog
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.*
 import com.bumptech.glide.Glide
-import org.ole.planet.myplanet.R
-import org.ole.planet.myplanet.databinding.ActivityLoginBinding
-import org.ole.planet.myplanet.datamanager.Service
-import org.ole.planet.myplanet.model.MyPlanet
-import org.ole.planet.myplanet.model.RealmMyTeam
-import org.ole.planet.myplanet.model.RealmUserModel
-import org.ole.planet.myplanet.model.User
+import io.realm.Realm
+import org.ole.planet.myplanet.*
+import org.ole.planet.myplanet.callback.SyncListener
+import org.ole.planet.myplanet.databinding.*
+import org.ole.planet.myplanet.datamanager.*
+import org.ole.planet.myplanet.model.*
 import org.ole.planet.myplanet.ui.community.HomeCommunityDialogFragment
 import org.ole.planet.myplanet.ui.feedback.FeedbackFragment
-import org.ole.planet.myplanet.ui.userprofile.TeamListAdapter
-import org.ole.planet.myplanet.utilities.FileUtils
-import org.ole.planet.myplanet.utilities.Utilities
+import org.ole.planet.myplanet.ui.userprofile.*
+import org.ole.planet.myplanet.utilities.*
+import org.ole.planet.myplanet.utilities.FileUtils.availableOverTotalMemoryFormattedString
+import org.ole.planet.myplanet.utilities.Utilities.getUrl
+import org.ole.planet.myplanet.utilities.Utilities.toast
+import java.text.Normalizer
+import java.util.regex.Pattern
 
 class LoginActivity : SyncActivity(), TeamListAdapter.OnItemClickListener {
     private lateinit var activityLoginBinding: ActivityLoginBinding
@@ -29,6 +36,7 @@ class LoginActivity : SyncActivity(), TeamListAdapter.OnItemClickListener {
     private var mAdapter: TeamListAdapter? = null
     private var backPressedTime: Long = 0
     private val BACK_PRESSED_INTERVAL: Long = 2000
+    private var fallbackLanguage: String = "en"
 
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -36,36 +44,24 @@ class LoginActivity : SyncActivity(), TeamListAdapter.OnItemClickListener {
         activityLoginBinding = ActivityLoginBinding.inflate(layoutInflater)
         setContentView(activityLoginBinding.root)
         lblLastSyncDate = activityLoginBinding.lblLastSyncDate
-        inputName = activityLoginBinding.inputName
-        inputPassword = activityLoginBinding.inputPassword
         btnSignIn = activityLoginBinding.btnSignin
         syncIcon = activityLoginBinding.syncIcon
-        becomeMember = activityLoginBinding.becomeMember
-        btnGuestLogin = activityLoginBinding.btnGuestLogin
-        imgBtnSetting = activityLoginBinding.imgBtnSetting
-        lblVersion = activityLoginBinding.lblVersion
-        btnLang = activityLoginBinding.btnLang
-        tvAvailableSpace = activityLoginBinding.tvAvailableSpace
-        openCommunity = activityLoginBinding.openCommunity
-        btnFeedback = activityLoginBinding.btnFeedback
-        customDeviceName = activityLoginBinding.customDeviceName
 
         service = Service(this)
 
-        tvAvailableSpace.text = FileUtils.availableOverTotalMemoryFormattedString
+        activityLoginBinding.tvAvailableSpace.text = availableOverTotalMemoryFormattedString
         changeLogoColor()
         declareElements()
         declareMoreElements()
         showWifiDialog()
         registerReceiver()
         forceSync = intent.getBooleanExtra("forceSync", false)
-        processedUrl = Utilities.getUrl()
+        processedUrl = getUrl()
         if (forceSync) {
             isSync = false
         }
         if (intent.hasExtra("versionInfo")) {
-            onUpdateAvailable(
-                (intent.getSerializableExtra("versionInfo") as MyPlanet?),
+            onUpdateAvailable((intent.getSerializableExtra("versionInfo") as MyPlanet?),
                 intent.getBooleanExtra("cancelable", false)
             )
         } else {
@@ -74,16 +70,16 @@ class LoginActivity : SyncActivity(), TeamListAdapter.OnItemClickListener {
         checkUsagesPermission()
         forceSyncTrigger()
 
-        if (Utilities.getUrl().isNotEmpty()) {
-            openCommunity.visibility = View.VISIBLE
-            openCommunity.setOnClickListener {
+        if (getUrl().isNotEmpty()) {
+            activityLoginBinding.openCommunity.visibility = View.VISIBLE
+            activityLoginBinding.openCommunity.setOnClickListener {
                 HomeCommunityDialogFragment().show(supportFragmentManager, "")
             }
             HomeCommunityDialogFragment().show(supportFragmentManager, "")
         } else {
-            openCommunity.visibility = View.GONE
+            activityLoginBinding.openCommunity.visibility = View.GONE
         }
-        btnFeedback.setOnClickListener {
+        activityLoginBinding.btnFeedback.setOnClickListener {
             FeedbackFragment().show(supportFragmentManager, "")
         }
 
@@ -100,11 +96,147 @@ class LoginActivity : SyncActivity(), TeamListAdapter.OnItemClickListener {
                 if (System.currentTimeMillis() - backPressedTime < BACK_PRESSED_INTERVAL) {
                     finish()
                 } else {
-                    Utilities.toast(this@LoginActivity, getString(R.string.press_back_again_to_exit))
+                    toast(this@LoginActivity, getString(R.string.press_back_again_to_exit))
                     backPressedTime = System.currentTimeMillis()
                 }
             }
         })
+    }
+
+    private fun declareElements() {
+        if (!defaultPref.contains("beta_addImageToMessage")) {
+            defaultPref.edit().putBoolean("beta_addImageToMessage", true).apply()
+        }
+        activityLoginBinding.customDeviceName.text = getCustomDeviceName()
+        activityLoginBinding.btnSignin.setOnClickListener {
+            if (TextUtils.isEmpty(activityLoginBinding.inputName.text.toString())) {
+                activityLoginBinding.inputName.error = getString(R.string.err_msg_name)
+            } else if (TextUtils.isEmpty(activityLoginBinding.inputPassword.text.toString())) {
+                activityLoginBinding.inputPassword.error = getString(R.string.err_msg_password)
+            } else {
+                if (mRealm.isClosed) {
+                    mRealm = Realm.getDefaultInstance()
+                }
+                val user = mRealm.where(RealmUserModel::class.java).equalTo("name", activityLoginBinding.inputName.text.toString()).findFirst()
+                if (user == null || !user.isArchived) {
+                    submitForm(activityLoginBinding.inputName.text.toString(), activityLoginBinding.inputPassword.text.toString())
+                } else {
+                    val builder = AlertDialog.Builder(this)
+                    builder.setMessage("member " + activityLoginBinding.inputName.text.toString() + " is archived")
+                    builder.setCancelable(false)
+                    builder.setPositiveButton("Ok") { dialog: DialogInterface, _: Int ->
+                        dialog.dismiss()
+                        activityLoginBinding.inputName.setText("")
+                        activityLoginBinding.inputPassword.setText("")
+                    }
+                    val dialog = builder.create()
+                    dialog.show()
+                }
+            }
+        }
+        if (!settings.contains("serverProtocol")) settings.edit().putString("serverProtocol", "http://").apply()
+        activityLoginBinding.becomeMember.setOnClickListener {
+            activityLoginBinding.inputName.setText("")
+            becomeAMember()
+        }
+
+        activityLoginBinding.imgBtnSetting.setOnClickListener {
+            activityLoginBinding.inputName.setText("")
+            settingDialog()
+        }
+
+        activityLoginBinding.btnGuestLogin.setOnClickListener {
+            if (getUrl().isNotEmpty()) {
+                activityLoginBinding.inputName.setText("")
+                showGuestLoginDialog()
+            } else {
+                toast(this, getString(R.string.please_enter_server_url_first))
+                settingDialog()
+            }
+        }
+    }
+
+    private fun declareMoreElements() {
+        try {
+            mRealm = Realm.getDefaultInstance()
+            syncIcon.setImageDrawable(ContextCompat.getDrawable(this, R.drawable.login_file_upload_animation))
+            syncIcon.scaleType
+            syncIconDrawable = syncIcon.drawable as AnimationDrawable
+            syncIcon.setOnClickListener {
+                syncIconDrawable.start()
+                isSync = false
+                forceSync = true
+                service.checkVersion(this, settings) }
+            declareHideKeyboardElements()
+            activityLoginBinding.lblVersion.text = "${resources.getText(R.string.version)} ${resources.getText(R.string.app_version)}"
+            activityLoginBinding.inputName.addTextChangedListener(MyTextWatcher(activityLoginBinding.inputName))
+            activityLoginBinding.inputPassword.addTextChangedListener(MyTextWatcher(activityLoginBinding.inputPassword))
+            activityLoginBinding.inputPassword.setOnEditorActionListener { _: TextView?, actionId: Int, event: KeyEvent? ->
+                if (actionId == EditorInfo.IME_ACTION_DONE || event != null && event.action == KeyEvent.ACTION_DOWN && event.keyCode == KeyEvent.KEYCODE_ENTER) {
+                    activityLoginBinding.btnSignin.performClick()
+                    return@setOnEditorActionListener true
+                }
+                false
+            }
+            setUpLanguageButton()
+            if (defaultPref.getBoolean("saveUsernameAndPassword", false)) {
+                activityLoginBinding.inputName.setText(settings.getString(getString(R.string.login_user), ""))
+                activityLoginBinding.inputPassword.setText(settings.getString(getString(R.string.login_password), ""))
+            }
+            if (NetworkUtils.isNetworkConnected) {
+                service.syncPlanetServers(mRealm) { success: String? ->
+                    toast(this, success)
+                }
+            }
+            activityLoginBinding.inputName.addTextChangedListener(object : TextWatcher {
+                override fun beforeTextChanged(s: CharSequence, start: Int, count: Int, after: Int) {}
+
+                override fun onTextChanged(s: CharSequence, start: Int, before: Int, count: Int) {
+                    val lowercaseText = s.toString().lowercase()
+                    if (s.toString() != lowercaseText) {
+                        activityLoginBinding.inputName.setText(lowercaseText)
+                        activityLoginBinding.inputName.setSelection(lowercaseText.length)
+                    }
+                }
+
+                override fun afterTextChanged(s: Editable) {}
+            })
+        } finally {
+            if (!mRealm.isClosed) {
+                mRealm.close()
+            }
+        }
+    }
+
+    private fun setUpLanguageButton() {
+        val languageKey = resources.getStringArray(R.array.language_keys)
+        val languages = resources.getStringArray(R.array.language)
+        val selectedLanguageKey = settings.getString(Constants.SELECTED_LANGUAGE, fallbackLanguage)
+        val index = languageKey.indexOf(selectedLanguageKey)
+        activityLoginBinding.btnLang.text = languages[index]
+        activityLoginBinding.btnLang.setOnClickListener {
+            AlertDialog.Builder(this)
+                .setTitle(R.string.select_language)
+                .setSingleChoiceItems(languages, index) { dialog, which ->
+                    val selectLanguageKey = languageKey[which]
+                    if (selectLanguageKey != settings.getString(Constants.SELECTED_LANGUAGE, fallbackLanguage)) {
+                        LocaleHelper.setLocale(this, selectLanguageKey)
+                        settings.edit().putString(Constants.SELECTED_LANGUAGE, selectLanguageKey).apply()
+                        activityLoginBinding.btnLang.text = languages[which]
+                        recreate()
+                    }
+                    dialog.dismiss()
+                }
+                .setNegativeButton(R.string.cancel, null)
+                .show()
+        }
+    }
+
+    private fun declareHideKeyboardElements() {
+        findViewById<View>(R.id.constraintLayout).setOnTouchListener { view: View?, _: MotionEvent? ->
+            hideKeyboard(view)
+            false
+        }
     }
 
     fun getTeamMembers() {
@@ -149,12 +281,12 @@ class LoginActivity : SyncActivity(), TeamListAdapter.OnItemClickListener {
                 .error(R.drawable.profile)
                 .into(activityLoginBinding.userProfile)
 
-            inputName.setText(user.name)
+            activityLoginBinding.inputName.setText(user.name)
         } else {
             if (user.source == "guest"){
                 val model = RealmUserModel.createGuestUser(user.name, mRealm, settings)?.let { mRealm.copyFromRealm(it) }
                 if (model == null) {
-                    Utilities.toast(this, getString(R.string.unable_to_login))
+                    toast(this, getString(R.string.unable_to_login))
                 } else {
                     saveUserInfoPref(settings, "", model)
                     onLogin()
@@ -162,6 +294,289 @@ class LoginActivity : SyncActivity(), TeamListAdapter.OnItemClickListener {
             } else {
                 submitForm(user.name, user.password)
             }
+        }
+    }
+
+    private fun submitForm(name: String?, password: String?) {
+        if (forceSyncTrigger()) {
+            return
+        }
+        val editor = settings.edit()
+        editor.putString("loginUserName", name)
+        editor.putString("loginUserPassword", password)
+        val isLoggedIn = authenticateUser(settings, name, password, false)
+        if (isLoggedIn) {
+            Toast.makeText(applicationContext, getString(R.string.thank_you), Toast.LENGTH_SHORT).show()
+            onLogin()
+            saveUsers(activityLoginBinding.inputName.text.toString(), activityLoginBinding.inputPassword.text.toString(), "member")
+        } else {
+            ManagerSync.instance?.login(name, password, object : SyncListener {
+                override fun onSyncStarted() {
+                    customProgressDialog?.setText(getString(R.string.please_wait))
+                    customProgressDialog?.show()
+                }
+
+                override fun onSyncComplete() {
+                    customProgressDialog?.dismiss()
+                    val log = authenticateUser(settings, name, password, true)
+                    if (log) {
+                        Toast.makeText(applicationContext, getString(R.string.thank_you), Toast.LENGTH_SHORT).show()
+                        onLogin()
+                        saveUsers(activityLoginBinding.inputName.text.toString(), activityLoginBinding.inputPassword.text.toString(), "member")
+                    } else {
+                        alertDialogOkay(getString(R.string.err_msg_login))
+                    }
+                    syncIconDrawable.stop()
+                    syncIconDrawable.selectDrawable(0)
+                }
+
+                override fun onSyncFailed(msg: String?) {
+                    toast(MainApplication.context, msg)
+                    customProgressDialog?.dismiss()
+                    syncIconDrawable.stop()
+                    syncIconDrawable.selectDrawable(0)
+                }
+            })
+        }
+        editor.apply()
+    }
+
+    private fun showGuestLoginDialog() {
+        try {
+            mRealm = Realm.getDefaultInstance()
+            mRealm.refresh()
+            val alertGuestLoginBinding = AlertGuestLoginBinding.inflate(LayoutInflater.from(this))
+            val v: View = alertGuestLoginBinding.root
+            alertGuestLoginBinding.etUserName.addTextChangedListener(object : TextWatcher {
+                override fun beforeTextChanged(s: CharSequence, start: Int, count: Int, after: Int) {}
+
+                override fun onTextChanged(s: CharSequence, start: Int, before: Int, count: Int) {
+                    val input = s.toString()
+                    val firstChar = if (input.isNotEmpty()) {
+                        input[0]
+                    } else {
+                        '\u0000'
+                    }
+                    var hasInvalidCharacters = false
+                    val hasSpecialCharacters: Boolean
+                    var hasDiacriticCharacters = false
+                    val normalizedText = Normalizer.normalize(s, Normalizer.Form.NFD)
+                    for (element in input) {
+                        if (element != '_' && element != '.' && element != '-' && !Character.isDigit(element) && !Character.isLetter(element)) {
+                            hasInvalidCharacters = true
+                            break
+                        }
+                    }
+                    val regex = ".*[ßäöüéèêæÆœøØ¿àìòùÀÈÌÒÙáíóúýÁÉÍÓÚÝâîôûÂÊÎÔÛãñõÃÑÕëïÿÄËÏÖÜŸåÅŒçÇðÐ].*"
+                    val pattern = Pattern.compile(regex)
+                    val matcher = pattern.matcher(input)
+                    hasSpecialCharacters = matcher.matches()
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                        hasDiacriticCharacters = !normalizedText.codePoints().allMatch { codePoint: Int -> Character.isLetterOrDigit(codePoint) || codePoint == '.'.code || codePoint == '-'.code || codePoint == '_'.code }
+                    }
+                    if (!Character.isDigit(firstChar) && !Character.isLetter(firstChar)) {
+                        alertGuestLoginBinding.etUserName.error = getString(R.string.must_start_with_letter_or_number)
+                    } else if (hasInvalidCharacters || hasDiacriticCharacters || hasSpecialCharacters) {
+                        alertGuestLoginBinding.etUserName.error = getString(R.string.only_letters_numbers_and_are_allowed)
+                    } else {
+                        val lowercaseText = input.lowercase()
+                        if (input != lowercaseText) {
+                            alertGuestLoginBinding.etUserName.setText(lowercaseText)
+                            alertGuestLoginBinding.etUserName.setSelection(lowercaseText.length)
+                        }
+                        alertGuestLoginBinding.etUserName.error = null
+                    }
+                }
+
+                override fun afterTextChanged(s: Editable) {}
+            })
+            val builder = AlertDialog.Builder(this)
+            builder.setTitle("Login As Guest")
+                .setView(v)
+                .setPositiveButton("Login", null)
+                .setNegativeButton("Cancel", null)
+            val dialog = builder.create()
+            dialog.show()
+            val login = dialog.getButton(AlertDialog.BUTTON_POSITIVE)
+            val cancel = dialog.getButton(AlertDialog.BUTTON_NEGATIVE)
+            login.setOnClickListener {
+                if (mRealm.isClosed) {
+                    mRealm = Realm.getDefaultInstance()
+                }
+                val username = alertGuestLoginBinding.etUserName.text.toString().trim { it <= ' ' }
+                val firstChar = if (username.isEmpty()) null else username[0]
+                var hasInvalidCharacters = false
+                var hasDiacriticCharacters = false
+                var hasSpecialCharacters = false
+                var isValid = true
+                val normalizedText = Normalizer.normalize(username, Normalizer.Form.NFD)
+                val regex = ".*[ßäöüéèêæÆœøØ¿àìòùÀÈÌÒÙáíóúýÁÉÍÓÚÝâîôûÂÊÎÔÛãñõÃÑÕëïÿÄËÏÖÜŸåÅŒçÇðÐ].*"
+                val pattern = Pattern.compile(regex)
+                val matcher = pattern.matcher(username)
+                if (TextUtils.isEmpty(username)) {
+                    alertGuestLoginBinding.etUserName.error = getString(R.string.username_cannot_be_empty)
+                    isValid = false
+                }
+                if (firstChar != null && !Character.isDigit(firstChar) && !Character.isLetter(firstChar)) {
+                    alertGuestLoginBinding.etUserName.error = getString(R.string.must_start_with_letter_or_number)
+                    isValid = false
+                } else {
+                    for (c in username.toCharArray()) {
+                        if (c != '_' && c != '.' && c != '-' && !Character.isDigit(c) && !Character.isLetter(c)) {
+                            hasInvalidCharacters = true
+                            break
+                        }
+                        hasSpecialCharacters = matcher.matches()
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                            hasDiacriticCharacters = !normalizedText.codePoints().allMatch {
+                                    codePoint -> Character.isLetterOrDigit(codePoint) || codePoint == '.'.code || codePoint == '-'.code || codePoint == '_'.code
+                            }
+                        }
+                    }
+                    if (hasInvalidCharacters || hasDiacriticCharacters || hasSpecialCharacters) {
+                        alertGuestLoginBinding.etUserName.error = getString(R.string.only_letters_numbers_and_are_allowed)
+                        isValid = false
+                    }
+                }
+                if (isValid) {
+                    val existingUser = mRealm.where(RealmUserModel::class.java).equalTo("name", username).findFirst()
+                    dialog.dismiss()
+                    if (existingUser != null) {
+                        if (existingUser._id?.contains("guest") == true) {
+                            showGuestDialog(username)
+                        } else if (existingUser._id?.contains("org.couchdb.user:") == true) {
+                            showUserAlreadyMemberDialog(username)
+                        }
+                    } else {
+                        val model = RealmUserModel.createGuestUser(username, mRealm, settings)
+                            ?.let { it1 ->
+                            mRealm.copyFromRealm(it1)
+                        }
+                        if (model == null) {
+                            toast(this, getString(R.string.unable_to_login))
+                        } else {
+                            saveUsers(username, "", "guest")
+                            saveUserInfoPref(settings, "", model)
+                            onLogin()
+                        }
+                    }
+                }
+            }
+            cancel.setOnClickListener { dialog.dismiss() }
+        } finally {
+            if (!mRealm.isClosed) {
+                mRealm.close()
+            }
+        }
+    }
+
+    private fun showGuestDialog(username: String) {
+        val builder = AlertDialog.Builder(this)
+        builder.setTitle("$username is already a guest")
+        builder.setMessage("Continue only if this is you")
+        builder.setCancelable(false)
+        builder.setNegativeButton("cancel") { dialog: DialogInterface, _: Int -> dialog.dismiss() }
+        builder.setPositiveButton("continue") { dialog: DialogInterface, _: Int ->
+            dialog.dismiss()
+            val model = RealmUserModel.createGuestUser(username, mRealm, settings)?.let { mRealm.copyFromRealm(it) }
+            if (model == null) {
+                toast(this, getString(R.string.unable_to_login))
+            } else {
+                saveUserInfoPref(settings, "", model)
+                onLogin()
+            }
+        }
+        val dialog = builder.create()
+        dialog.show()
+    }
+
+    private fun showUserAlreadyMemberDialog(username: String) {
+        val builder = AlertDialog.Builder(this)
+        builder.setTitle("$username is already a member")
+        builder.setMessage("Continue to login if this is you")
+        builder.setCancelable(false)
+        builder.setNegativeButton("Cancel") { dialog: DialogInterface, _: Int -> dialog.dismiss() }
+        builder.setPositiveButton("login") { dialog: DialogInterface, _: Int ->
+            dialog.dismiss()
+            activityLoginBinding.inputName.setText(username)
+        }
+        val dialog = builder.create()
+        dialog.show()
+    }
+
+    fun saveUsers(name: String?, password: String?, source: String) {
+        if (source === "guest") {
+            val newUser = User("", name, password, "", "guest")
+            val existingUsers: MutableList<User> = ArrayList(
+                prefData.getSAVEDUSERS()
+            )
+            var newUserExists = false
+            for ((_, name1) in existingUsers) {
+                if (name1 == newUser.name?.trim { it <= ' ' }) {
+                    newUserExists = true
+                    break
+                }
+            }
+            if (!newUserExists) {
+                existingUsers.add(newUser)
+                prefData.setSAVEDUSERS(existingUsers)
+            }
+        } else if (source === "member") {
+            var userProfile = profileDbHandler.userModel?.userImage
+            var fullName: String? = profileDbHandler.userModel?.getFullName()
+            if (userProfile == null) {
+                userProfile = ""
+            }
+            if (fullName?.trim { it <= ' ' }?.isEmpty() == true) {
+                fullName = profileDbHandler.userModel?.name
+            }
+            val newUser = User(fullName, name, password, userProfile, "member")
+            val existingUsers: MutableList<User> = ArrayList(prefData.getSAVEDUSERS())
+            var newUserExists = false
+            for ((fullName1) in existingUsers) {
+                if (fullName1 == newUser.fullName?.trim { it <= ' ' }) {
+                    newUserExists = true
+                    break
+                }
+            }
+            if (!newUserExists) {
+                existingUsers.add(newUser)
+                prefData.setSAVEDUSERS(existingUsers)
+            }
+        }
+    }
+
+    private fun becomeAMember() {
+        if (getUrl().isNotEmpty()) {
+            startActivity(Intent(this, BecomeMemberActivity::class.java))
+        } else {
+            toast(this, getString(R.string.please_enter_server_url_first))
+            settingDialog()
+        }
+    }
+
+    fun getCustomDeviceName(): String? {
+        return settings.getString("customDeviceName", NetworkUtils.getDeviceName())
+    }
+
+    private fun resetGuestAsMember(username: String?) {
+        val existingUsers = prefData.getSAVEDUSERS().toMutableList()
+        var newUserExists = false
+        for ((_, name) in existingUsers) {
+            if (name == username) {
+                newUserExists = true
+                break
+            }
+        }
+        if (newUserExists) {
+            val iterator = existingUsers.iterator()
+            while (iterator.hasNext()) {
+                val (_, name) = iterator.next()
+                if (name == username) {
+                    iterator.remove()
+                }
+            }
+            prefData.setSAVEDUSERS(existingUsers)
         }
     }
